refactor(structure): reorganize utils and integrators directories
Move utility headers and source files (conditional_o_stream_wrapper,
mesh_generator, progress_bar) and time integrator headers/source files
into dedicated `utils` and `time_integrator` directories.

Update all include paths accordingly and adjust CMakeLists.txt
target to reference the new file locations.

Rename and relocate files consistently across the project.

Refs: #5
