+++ include/utils/conditional_o_stream_wrapper.hpp
